+++ test/utils.py
@@ -49,10 +49,8 @@ def get_pool(client: MongoClient) -> Pool:
 
 def get_pools(client: MongoClient) -> list[Pool]:
     """Get all pools."""
-    return [
-        server.pool
-        for server in (client._get_topology()).select_servers(any_server_selector, _Op.TEST)
-    ]
+    servers = (client._get_topology()).select_servers(any_server_selector, _Op.TEST)
+    return [server.pool for server in servers]
 
 
 def wait_until(predicate, success_description, timeout=10):